Resolve the iOS storage budget as an Int, not an Int64
HARD27-028.

`getStorageBudget` resolved `availableBytes` as an `Int64`, which is the type of
`volumeAvailableCapacityForImportantUsage`. Every other numeric this app resolves
is an `Int` — `data.count` in `IOSFtp`, the FTP entry `size` — and `Int` is what
Capacitor coerces to a JS number without a bridging cast. An `Int64` reaching
`coerceDictionaryToJSObject` relies on `NSNumber` bridging that nothing else in
this codebase depends on. iOS is 64-bit only, so `Int` holds any byte count a
volume can report.

The mirror in `HvscStorageBudget` follows: it takes the important-usage figure as
`Int64` and the plain capacity as `Int`, matching the two Foundation types, and
returns `Int`. `HvscStorageBudgetTests` gains a case that a 60 GiB `Int64` figure
survives the narrowing, and the source assertion now requires the plugin to
narrow it explicitly.

Tests: `swift test` in `ios/native-tests` — 86 tests, zero failures.
`docs/internals/ios-parity-matrix.md` test count updated to match.

This is a change to Swift that no runner has compiled. iOS run 33812380218 covers
the previous commit; dispatch again on this one once it finishes, since a
dispatch cancels a run in progress.

+++ ios/native-tests/Sources/NativeValidation/HvscStorageBudget.swift
@@ -27,9 +27,9 @@ public enum HvscStorageBudget {
     /// one costs a retry.
     public static func availableBytes(
         importantUsageCapacity: Int64?,
-        volumeAvailableCapacity: Int64?
-    ) -> Int64 {
-        if let important = importantUsageCapacity, important > 0 { return important }
+        volumeAvailableCapacity: Int?
+    ) -> Int {
+        if let important = importantUsageCapacity, important > 0 { return Int(important) }
         if let available = volumeAvailableCapacity, available > 0 { return available }
         return 0
     }
